Remove all the gfx hacks for cross platform issues
This came to me after being annoyed with all the changes I made
yesterday to compensate. I knew it had to be more simple.

Part of it was the flip_vert_y setting in the shader. This was flipping
the vertices of the quad, not the actual model vertices.

Also this lets us fix flipspace for metal (ZO instead of NO).

Now we can undo all that in favor of face_winding per API. This is
normal.

All the other stuff was just piled on hacks to fix the initial incorrect
fix.