Large refactor of event/scenario parsing and caching
We now cache in the io module instead of the gui module.
We pass in a buffer to parse the scenario/event instead of opening a
file each time. This makes it closer to the other read functions like
read_mesh, read_texture, etc.

Scenario and event now store there own data as well.

This removes dependency between modules.