Merge map_data_t and map_t into map_t
These didn't need to be separated. map_t only held the current
renderable mesh and texture. Now we just pass that to the model_t and
create what we need. We don't need to hold on to it.

But we do need the lighting so the model_t now has a
lighting (lighting_t) field.